File Structure, add TACU_isAP to AP units
