+++ addons/mercenaries/CfgVehicles.hpp
@@ -6,14 +6,14 @@ class CfgVehicles {
     class TACU_Main_U_INDEP_Soldier_Base;
 
     // Units - Opfor
-    #include "CfgVehicles_O_Low.hpp"
-    #include "CfgVehicles_O_Medium.hpp"
-    #include "CfgVehicles_O_High.hpp"
+    #include "opfor\CfgVehicles_O_Low.hpp"
+    #include "opfor\CfgVehicles_O_Medium.hpp"
+    #include "opfor\CfgVehicles_O_High.hpp"
 
     // Units - Independent
-    #include "CfgVehicles_I_Low.hpp"
-    #include "CfgVehicles_I_Medium.hpp"
-    #include "CfgVehicles_I_High.hpp"
+    #include "independent\CfgVehicles_I_Low.hpp"
+    #include "independent\CfgVehicles_I_Medium.hpp"
+    #include "independent\CfgVehicles_I_High.hpp"
 
     #include "CfgVehicles_Backpacks.hpp"
 };
