--- addons/mercenaries/CfgVehicles_O_Medium.hpp
@@ -1,176 +1,0 @@
-class TACU_Mercenaries_U_O_Medium_Rifleman: TACU_Mercenaries_U_O_Low_Rifleman {
-    MACRO_OPFOR_MEDIUM_GEAR;
-    weapons[] = {
-        "TACU_Mercenaries_W_HK416_Medium",
-        "hgun_Pistol_heavy_01_F",
-        "Throw", "Put"
-    };
-    respawnWeapons[] = {
-        "TACU_Mercenaries_W_HK416_Medium",
-        "hgun_Pistol_heavy_01_F",
-        "Throw", "Put"
-    };
-    //EDITORPREVIEW(TACU_Mercenaries_U_O_Medium_Rifleman);
-};
-class TACU_Mercenaries_U_O_Medium_AT: TACU_Mercenaries_U_O_Low_AT {
-    MACRO_OPFOR_MEDIUM_GEAR;
-    weapons[] = {
-        "launch_RPG32_tan_lxWS",
-        "TACU_Mercenaries_W_HK416_Medium",
-        "hgun_Pistol_heavy_01_F",
-        "Throw", "Put"
-    };
-    respawnWeapons[] = {
-        "launch_RPG32_tan_lxWS",
-        "TACU_Mercenaries_W_HK416_Medium",
-        "hgun_Pistol_heavy_01_F",
-        "Throw", "Put"
-    };
-    //EDITORPREVIEW(TACU_Mercenaries_U_O_Medium_AT);
-};
-class TACU_Mercenaries_U_O_Medium_AP: TACU_Mercenaries_U_O_Low_AP {
-    MACRO_OPFOR_MEDIUM_GEAR;
-    weapons[] = {
-        "launch_RPG32_tan_lxWS",
-        "TACU_Mercenaries_W_HK416_Medium",
-        "hgun_Pistol_heavy_01_F",
-        "Throw", "Put"
-    };
-    respawnWeapons[] = {
-        "launch_RPG32_tan_lxWS",
-        "TACU_Mercenaries_W_HK416_Medium",
-        "hgun_Pistol_heavy_01_F",
-        "Throw", "Put"
-    };
-    //EDITORPREVIEW(TACU_Mercenaries_U_O_Medium_AP);
-};
-class TACU_Mercenaries_U_O_Medium_MG: TACU_Mercenaries_U_O_Low_MG {
-    MACRO_OPFOR_MEDIUM_GEAR;
-    weapons[] = {
-        "TACU_Mercenaries_W_M249_Medium",
-        "hgun_Pistol_heavy_01_F",
-        "Throw", "Put"
-    };
-    respawnWeapons[] = {
-        "TACU_Mercenaries_W_M249_Medium",
-        "hgun_Pistol_heavy_01_F",
-        "Throw", "Put"
-    };
-    //EDITORPREVIEW(TACU_Mercenaries_U_O_Medium_MG);
-};
-class TACU_Mercenaries_U_O_Medium_Medic: TACU_Mercenaries_U_O_Low_Medic {
-    MACRO_OPFOR_MEDIUM_GEAR;
-    weapons[] = {
-        "TACU_Mercenaries_W_HK416_Medium",
-        "hgun_Pistol_heavy_01_F",
-        "Throw", "Put"
-    };
-    respawnWeapons[] = {
-        "TACU_Mercenaries_W_HK416_Medium",
-        "hgun_Pistol_heavy_01_F",
-        "Throw", "Put"
-    };
-    //EDITORPREVIEW(TACU_Mercenaries_U_O_Medium_Medic);
-};
-class TACU_Mercenaries_U_O_Medium_Engineer: TACU_Mercenaries_U_O_Low_Engineer {
-    MACRO_OPFOR_MEDIUM_GEAR;
-    weapons[] = {
-        "TACU_Mercenaries_W_HK416_Medium",
-        "hgun_Pistol_heavy_01_F",
-        "Throw", "Put"
-    };
-    respawnWeapons[] = {
-        "TACU_Mercenaries_W_HK416_Medium",
-        "hgun_Pistol_heavy_01_F",
-        "Throw", "Put"
-    };
-    //EDITORPREVIEW(TACU_Mercenaries_U_O_Medium_Engineer);
-};
-class TACU_Mercenaries_U_O_Medium_Demolitions: TACU_Mercenaries_U_O_Low_Demolitions {
-    MACRO_OPFOR_MEDIUM_GEAR;
-    weapons[] = {
-        "TACU_Mercenaries_W_HK416_Medium",
-        "hgun_Pistol_heavy_01_F",
-        "Throw", "Put"
-    };
-    respawnWeapons[] = {
-        "TACU_Mercenaries_W_HK416_Medium",
-        "hgun_Pistol_heavy_01_F",
-        "Throw", "Put"
-    };
-    //EDITORPREVIEW(TACU_Mercenaries_U_O_Medium_Demolitions);
-};
-class TACU_Mercenaries_U_O_Medium_Grenadier: TACU_Mercenaries_U_O_Low_Grenadier {
-    MACRO_OPFOR_MEDIUM_GEAR;
-    weapons[] = {
-        "TACU_Mercenaries_W_HK416_GL_Medium",
-        "hgun_Pistol_heavy_01_F",
-        "Throw", "Put"
-    };
-    respawnWeapons[] = {
-        "TACU_Mercenaries_W_HK416_GL_Medium",
-        "hgun_Pistol_heavy_01_F",
-        "Throw", "Put"
-    };
-    //EDITORPREVIEW(TACU_Mercenaries_U_O_Medium_Grenadier);
-};
-class TACU_Mercenaries_U_O_Medium_Marksman: TACU_Mercenaries_U_O_Low_Marksman {
-    MACRO_OPFOR_MEDIUM_GEAR;
-    weapons[] = {
-        "Binocular",
-        "TACU_Mercenaries_W_HK417_Medium",
-        "hgun_Pistol_heavy_01_F",
-        "Throw", "Put"
-    };
-    respawnWeapons[] = {
-        "Binocular",
-        "TACU_Mercenaries_W_HK417_Medium",
-        "hgun_Pistol_heavy_01_F",
-        "Throw", "Put"
-    };
-    //EDITORPREVIEW(TACU_Mercenaries_U_O_Low_Marksman);
-};
-class TACU_Mercenaries_U_O_Medium_TeamLeader: TACU_Mercenaries_U_O_Low_TeamLeader {
-    MACRO_OPFOR_MEDIUM_GEAR;
-    weapons[] = {
-        "TACU_Mercenaries_W_HK416A5_Medium",
-        "hgun_Pistol_heavy_01_F",
-        "Throw", "Put"
-    };
-    respawnWeapons[] = {
-        "TACU_Mercenaries_W_HK416A5_Medium",
-        "hgun_Pistol_heavy_01_F",
-        "Throw", "Put"
-    };
-    //EDITORPREVIEW(TACU_Mercenaries_U_O_Medium_TeamLeader);
-};
-class TACU_Mercenaries_U_O_Medium_SquadLeader: TACU_Mercenaries_U_O_Low_SquadLeader {
-    MACRO_OPFOR_MEDIUM_GEAR;
-    weapons[] = {
-        "Binocular",
-        "TACU_Mercenaries_W_HK416_GL_Medium",
-        "hgun_Pistol_heavy_01_F",
-        "Throw", "Put"
-    };
-    respawnWeapons[] = {
-        "Binocular",
-        "TACU_Mercenaries_W_HK416_GL_Medium",
-        "hgun_Pistol_heavy_01_F",
-        "Throw", "Put"
-    };
-    //EDITORPREVIEW(TACU_Mercenaries_U_O_Medium_SquadLeader);
-};
-class TACU_Mercenaries_U_O_Medium_Crewman: TACU_Mercenaries_U_O_Low_Crewman {
-    MACRO_OPFOR_MEDIUM_GEAR;
-    weapons[] = {
-        "TACU_Mercenaries_W_HK416_Medium",
-        "hgun_Pistol_heavy_01_F",
-        "Throw", "Put"
-    };
-    respawnWeapons[] = {
-        "TACU_Mercenaries_W_HK416_Medium",
-        "hgun_Pistol_heavy_01_F",
-        "Throw", "Put"
-    };
-    //EDITORPREVIEW(TACU_Mercenaries_U_O_Medium_Crewman);
-};
