+++ addons/mercenaries/independent/CfgVehicles_I_Low.hpp
@@ -1,33 +1,33 @@
-class TACU_Mercenaries_U_O_Low_Rifleman: TACU_Main_U_OPFOR_Soldier_Base {
+class TACU_Mercenaries_U_I_Low_Rifleman: TACU_Main_U_INDEP_Soldier_Base {
     author = "Mike";
     editorSubcategory = "TACU_Mercenaries_EdSubCat_Low";
     displayName = "Rifleman";
-    faction = "TACU_Mercenaries_O";
-    genericNames = "TACU_IranianNames";
+    faction = "TACU_Mercenaries_I";
+    genericNames = "CivMen";
     icon = "iconMan";
-    identityTypes[] = {"LanguagePER_F", "Head_TK"};
+    identityTypes[] = {"LanguageENGFRE_F", "Head_African", "Head_NATO"};
     role = "Rifleman";
     scope = 2;
     scopeCurator = 2;
-    uniformClass = "U_BG_Guerilla2_1";
+    uniformClass = "U_BG_Guerilla2_3";
     linkedItems[] = {
         DEFAULT_ITEMS_RADIO,
-        "V_Chestrig_khk",
-        "H_Bandanna_cbr"
+        "V_Chestrig_oli",
+        "H_Bandanna_khk_hs"
     };
     respawnLinkedItems[] = {
         DEFAULT_ITEMS_RADIO,
-        "V_Chestrig_khk",
-        "H_Bandanna_cbr"
+        "V_Chestrig_oli",
+        "H_Bandanna_khk_hs"
     };
     weapons[] = {
-        "TACU_Mercenaries_W_HK416A5_Low",
-        "hgun_Pistol_heavy_01_F",
+        "TACU_Mercenaries_W_Tar21_Low",
+        "hgun_Pistol_heavy_01_green_F",
         "Throw", "Put"
     };
     respawnWeapons[] = {
-        "TACU_Mercenaries_W_HK416A5_Low",
-        "hgun_Pistol_heavy_01_F",
+        "TACU_Mercenaries_W_Tar21_Low",
+        "hgun_Pistol_heavy_01_green_F",
         "Throw", "Put"
     };
     magazines[] = {
@@ -43,14 +43,17 @@ class TACU_Mercenaries_U_O_Low_Rifleman: TACU_Main_U_OPFOR_Soldier_Base {
         mag_2("SmokeShell")
     };
     headgearList[] = {
-        "H_Bandanna_cbr", 0.3,
-        "H_Watchcap_cbr", 0.3,
-        "H_Booniehat_khk_hs", 0.3,
-        "H_Booniehat_khk", 0.3,
-        "H_Booniehat_tan", 0.3,
-        "H_Cap_tan", 0.3,
-        "H_ShemagOpen_tan", 0.3,
-        "H_HelmetB_sand", 0.1
+        "H_Bandanna_khk", 0.3,
+        "H_Bandanna_khk_hs", 0.3,
+        "H_Bandanna_sgg", 0.3,
+        "H_Watchcap_camo", 0.3,
+        "H_Watchcap_khk", 0.3,
+        "H_Cap_grn", 0.3,
+        "H_Cap_oli", 0.3,
+        "H_Cap_oli_hs", 0.3,
+        "H_Shemag_olive", 0.3,
+        "H_Shemag_olive_hs", 0.3,
+        "H_HelmetB", 0.1
     };
     CBA_facewearList[] = {
         "", 0.65,
@@ -61,155 +64,156 @@ class TACU_Mercenaries_U_O_Low_Rifleman: TACU_Main_U_OPFOR_Soldier_Base {
         "G_Bandanna_aviator", 0.25,
         "G_Spectacles_Tinted", 0.25
     };
-    //EDITORPREVIEW(TACU_Mercenaries_U_O_Low_Rifleman);
+    //EDITORPREVIEW(TACU_Mercenaries_U_I_Low_Rifleman);
 };
-class TACU_Mercenaries_U_O_Low_AT: TACU_Mercenaries_U_O_Low_Rifleman {
+class TACU_Mercenaries_U_I_Low_AT: TACU_Mercenaries_U_I_Low_Rifleman {
     displayName = "Rifleman (AT)";
     role = "MissileSpecialist";
     icon = "iconManAT";
-    backpack = "TACU_Mercenaries_B_AT_FieldPack_Sand";
+    backpack = "TACU_Mercenaries_B_AT_FieldPack_Green";
     weapons[] = {
-        "launch_RPG32_tan_lxWS",
-        "TACU_Mercenaries_W_HK416_Low",
-        "hgun_Pistol_heavy_01_F",
+        "launch_MRAWS_green_rail_F",
+        "TACU_Mercenaries_W_Tar21_Low",
+        "hgun_Pistol_heavy_01_green_F",
         "Throw", "Put"
     };
     respawnWeapons[] = {
-        "launch_RPG32_tan_lxWS",
-        "TACU_Mercenaries_W_HK416_Low",
-        "hgun_Pistol_heavy_01_F",
+        "launch_MRAWS_green_rail_F",
+        "TACU_Mercenaries_W_Tar21_Low",
+        "hgun_Pistol_heavy_01_green_F",
         "Throw", "Put"
     };
     magazines[] = {
-        "RPG32_F",
+        "MRAWS_HEAT_F",
         mag_8("TACU_Magazine_30Rnd_STANAG"),
         mag_2("TACU_Magazine_11Rnd_FNX"),
         mag_2("HandGrenade"),
         mag_2("SmokeShell")
     };
     respawnMagazines[] = {
-        "RPG32_F",
+        "MRAWS_HEAT_F",
         mag_8("TACU_Magazine_30Rnd_STANAG"),
         mag_2("TACU_Magazine_11Rnd_FNX"),
         mag_2("HandGrenade"),
         mag_2("SmokeShell")
     };
-    //EDITORPREVIEW(TACU_Mercenaries_U_O_Low_AT);
+    //EDITORPREVIEW(TACU_Mercenaries_U_I_Low_AT);
 };
-class TACU_Mercenaries_U_O_Low_AP: TACU_Mercenaries_U_O_Low_AT {
+class TACU_Mercenaries_U_I_Low_AP: TACU_Mercenaries_U_I_Low_AT {
     displayName = "Rifleman (AP)";
-    backpack = "TACU_Mercenaries_B_AP_FieldPack_Sand";
+    backpack = "TACU_Mercenaries_B_AP_FieldPack_Green";
+    TACU_isAP = 1;
     magazines[] = {
-        "RPG32_HE_F",
+        "MRAWS_HE_F",
         mag_8("TACU_Magazine_30Rnd_STANAG"),
         mag_2("TACU_Magazine_11Rnd_FNX"),
         mag_2("HandGrenade"),
         mag_2("SmokeShell")
     };
     respawnMagazines[] = {
-        "RPG32_HE_F",
+        "MRAWS_HE_F",
         mag_8("TACU_Magazine_30Rnd_STANAG"),
         mag_2("TACU_Magazine_11Rnd_FNX"),
         mag_2("HandGrenade"),
         mag_2("SmokeShell")
     };
-    //EDITORPREVIEW(TACU_Mercenaries_U_O_Low_AP);
+    //EDITORPREVIEW(TACU_Mercenaries_U_I_Low_AP);
 };
-class TACU_Mercenaries_U_O_Low_MG: TACU_Mercenaries_U_O_Low_Rifleman {
+class TACU_Mercenaries_U_I_Low_MG: TACU_Mercenaries_U_I_Low_Rifleman {
     displayName = "Machine Gunner";
-    backpack = "TACU_Mercenaries_B_MG_FieldPack_Sand";
+    backpack = "TACU_Mercenaries_B_MG_FieldPack_Green";
     icon = "iconManMG";
     role = "MachineGunner";
     weapons[] = {
-        "TACU_Mercenaries_W_M249_Low",
-        "hgun_Pistol_heavy_01_F",
+        "TACU_Mercenaries_W_Stoner_Low",
+        "hgun_Pistol_heavy_01_green_F",
         "Throw", "Put"
     };
     respawnWeapons[] = {
-        "TACU_Mercenaries_W_M249_Low",
-        "hgun_Pistol_heavy_01_F",
+        "TACU_Mercenaries_W_Stoner_Low",
+        "hgun_Pistol_heavy_01_green_F",
         "Throw", "Put"
     };
     magazines[] = {
-        "TACU_Magazine_200Rnd_556_Box",
+        "TACU_Magazine_200Rnd_65_Box",
         mag_2("TACU_Magazine_11Rnd_FNX"),
         mag_2("HandGrenade"),
         mag_2("SmokeShell")
     };
     respawnMagazines[] = {
-        "TACU_Magazine_200Rnd_556_Box",
+        "TACU_Magazine_200Rnd_65_Box",
         mag_2("TACU_Magazine_11Rnd_FNX"),
         mag_2("HandGrenade"),
         mag_2("SmokeShell")
     };
-    //EDITORPREVIEW(TACU_Mercenaries_U_O_Low_MG);
+    //EDITORPREVIEW(TACU_Mercenaries_U_I_Low_Rifleman_MG);
 };
-class TACU_Mercenaries_U_O_Low_Medic: TACU_Mercenaries_U_O_Low_Rifleman {
+class TACU_Mercenaries_U_I_Low_Medic: TACU_Mercenaries_U_I_Low_Rifleman {
     displayName = "Medic";
-    backpack = "TACU_Mercenaries_B_Medic_FieldPack_Sand";
+    backpack = "TACU_Mercenaries_B_Medic_FieldPack_Green";
     icon = "iconManMedic";
     role = "CombatLifeSaver";
     attendant = 1;
     weapons[] = {
-        "TACU_Mercenaries_W_HK416_Low",
-        "hgun_Pistol_heavy_01_F",
+        "TACU_Mercenaries_W_Ctar21_Low",
+        "hgun_Pistol_heavy_01_green_F",
         "Throw", "Put"
     };
     respawnWeapons[] = {
-        "TACU_Mercenaries_W_HK416_Low",
-        "hgun_Pistol_heavy_01_F",
+        "TACU_Mercenaries_W_Ctar21_Low",
+        "hgun_Pistol_heavy_01_green_F",
         "Throw", "Put"
     };
-    //EDITORPREVIEW(TACU_Mercenaries_U_O_Low_Medic);
+    //EDITORPREVIEW(TACU_Mercenaries_U_I_Low_Rifleman_Medic);
 };
-class TACU_Mercenaries_U_O_Low_Engineer: TACU_Mercenaries_U_O_Low_Rifleman {
+class TACU_Mercenaries_U_I_Low_Engineer: TACU_Mercenaries_U_I_Low_Rifleman {
     displayName = "Engineer";
     backpack = "TACU_Mercenaries_B_Engineer_FieldPack_Green";
     icon = "iconManEngineer";
     role = "Sapper";
     engineer = 1;
     weapons[] = {
-        "TACU_Mercenaries_W_HK416_Low",
-        "hgun_Pistol_heavy_01_F",
+        "TACU_Mercenaries_W_Ctar21_Low",
+        "hgun_Pistol_heavy_01_green_F",
         "Throw", "Put"
     };
     respawnWeapons[] = {
-        "TACU_Mercenaries_W_HK416_Low",
-        "hgun_Pistol_heavy_01_F",
+        "TACU_Mercenaries_W_Ctar21_Low",
+        "hgun_Pistol_heavy_01_green_F",
         "Throw", "Put"
     };
-    //EDITORPREVIEW(TACU_Mercenaries_U_O_Low_Engineer);
+    //EDITORPREVIEW(TACU_Mercenaries_U_I_Low_Rifleman_Engineer);
 };
-class TACU_Mercenaries_U_O_Low_Demolitions: TACU_Mercenaries_U_O_Low_Rifleman {
+class TACU_Mercenaries_U_I_Low_Demolitions: TACU_Mercenaries_U_I_Low_Rifleman {
     displayName = "Demolitions";
-    backpack = "TACU_Mercenaries_B_Demolitions_FieldPack_Sand";
+    backpack = "TACU_Mercenaries_B_Demolitions_FieldPack_Green";
     icon = "iconManExplosive";
     role = "Sapper";
     engineer = 1;
     canDeactivateMines = 1;
     weapons[] = {
-        "TACU_Mercenaries_W_HK416_Low",
-        "hgun_Pistol_heavy_01_F",
+        "TACU_Mercenaries_W_Ctar21_Low",
+        "hgun_Pistol_heavy_01_green_F",
         "Throw", "Put"
     };
     respawnWeapons[] = {
-        "TACU_Mercenaries_W_HK416_Low",
-        "hgun_Pistol_heavy_01_F",
+        "TACU_Mercenaries_W_Ctar21_Low",
+        "hgun_Pistol_heavy_01_green_F",
         "Throw", "Put"
     };
-    //EDITORPREVIEW(TACU_Mercenaries_U_O_Low_Demolitions);
+    //EDITORPREVIEW(TACU_Mercenaries_U_I_Low_Rifleman_Demolitions);
 };
-class TACU_Mercenaries_U_O_Low_Grenadier: TACU_Mercenaries_U_O_Low_Rifleman {
+class TACU_Mercenaries_U_I_Low_Grenadier: TACU_Mercenaries_U_I_Low_Rifleman {
     displayName = "Grenadier";
     role = "Grenadier";
     weapons[] = {
-        "TACU_Mercenaries_W_HK416_GL_Low",
-        "hgun_Pistol_heavy_01_F",
+        "TACU_Mercenaries_W_Gtar21_Low",
+        "hgun_Pistol_heavy_01_green_F",
         "Throw", "Put"
     };
     respawnWeapons[] = {
-        "TACU_Mercenaries_W_HK416_GL_Low",
-        "hgun_Pistol_heavy_01_F",
+        "TACU_Mercenaries_W_Gtar21_Low",
+        "hgun_Pistol_heavy_01_green_F",
         "Throw", "Put"
     };
     magazines[] = {
@@ -226,22 +230,22 @@ class TACU_Mercenaries_U_O_Low_Grenadier: TACU_Mercenaries_U_O_Low_Rifleman {
         mag_2("HandGrenade"),
         mag_2("SmokeShell")
     };
-    //EDITORPREVIEW(TACU_Mercenaries_U_O_Low_Grenadier);
+    //EDITORPREVIEW(TACU_Mercenaries_U_I_Low_Grenadier);
 };
-class TACU_Mercenaries_U_O_Low_Marksman: TACU_Mercenaries_U_O_Low_Rifleman {
+class TACU_Mercenaries_U_I_Low_Marksman: TACU_Mercenaries_U_I_Low_Rifleman {
     displayName = "Marksman";
     icon = "iconManRecon";
     role = "Marksman";
     weapons[] = {
         "Binocular",
-        "TACU_Mercenaries_W_HK417_Low",
-        "hgun_Pistol_heavy_01_F",
+        "TACU_Mercenaries_W_M14_Low",
+        "hgun_Pistol_heavy_01_green_F",
         "Throw", "Put"
     };
     respawnWeapons[] = {
         "Binocular",
-        "TACU_Mercenaries_W_HK417_Low",
-        "hgun_Pistol_heavy_01_F",
+        "TACU_Mercenaries_W_M14_Low",
+        "hgun_Pistol_heavy_01_green_F",
         "Throw", "Put"
     };
     magazines[] = {
@@ -256,36 +260,36 @@ class TACU_Mercenaries_U_O_Low_Marksman: TACU_Mercenaries_U_O_Low_Rifleman {
         mag_2("HandGrenade"),
         mag_2("SmokeShell")
     };
-    //EDITORPREVIEW(TACU_Mercenaries_U_O_Low_Marksman);
+    //EDITORPREVIEW(TACU_Mercenaries_U_I_Low_Marksman);
 };
-class TACU_Mercenaries_U_O_Low_TeamLeader: TACU_Mercenaries_U_O_Low_Rifleman {
+class TACU_Mercenaries_U_I_Low_TeamLeader: TACU_Mercenaries_U_I_Low_Rifleman {
     displayName = "Team Leader";
     icon = "iconManLeader";
     weapons[] = {
-        "TACU_Mercenaries_W_HK416A5_Low",
-        "hgun_Pistol_heavy_01_F",
+        "TACU_Mercenaries_W_Tar21_Low",
+        "hgun_Pistol_heavy_01_green_F",
         "Throw", "Put"
     };
     respawnWeapons[] = {
-        "TACU_Mercenaries_W_HK416A5_Low",
-        "hgun_Pistol_heavy_01_F",
+        "TACU_Mercenaries_W_Tar21_Low",
+        "hgun_Pistol_heavy_01_green_F",
         "Throw", "Put"
     };
-    //EDITORPREVIEW(TACU_Mercenaries_U_O_Low_TeamLeader);
+    //EDITORPREVIEW(TACU_Mercenaries_U_I_Low_TeamLeader);
 };
-class TACU_Mercenaries_U_O_Low_SquadLeader: TACU_Mercenaries_U_O_Low_Rifleman {
+class TACU_Mercenaries_U_I_Low_SquadLeader: TACU_Mercenaries_U_I_Low_Rifleman {
     displayName = "Squad Leader";
     icon = "iconManLeader";
     weapons[] = {
         "Binocular",
-        "TACU_Mercenaries_W_HK416_GL_Low",
-        "hgun_Pistol_heavy_01_F",
+        "TACU_Mercenaries_W_Gtar21_Low",
+        "hgun_Pistol_heavy_01_green_F",
         "Throw", "Put"
     };
     respawnWeapons[] = {
         "Binocular",
-        "TACU_Mercenaries_W_HK416_GL_Low",
-        "hgun_Pistol_heavy_01_F",
+        "TACU_Mercenaries_W_Gtar21_Low",
+        "hgun_Pistol_heavy_01_green_F",
         "Throw", "Put"
     };
     magazines[] = {
@@ -302,22 +306,22 @@ class TACU_Mercenaries_U_O_Low_SquadLeader: TACU_Mercenaries_U_O_Low_Rifleman {
         mag_2("HandGrenade"),
         mag_2("SmokeShell")
     };
-    //EDITORPREVIEW(TACU_Mercenaries_U_O_Low_SquadLeader);
+    //EDITORPREVIEW(TACU_Mercenaries_U_I_Low_SquadLeader);
 };
-class TACU_Mercenaries_U_O_Low_Crewman: TACU_Mercenaries_U_O_Low_Rifleman {
+class TACU_Mercenaries_U_I_Low_Crewman: TACU_Mercenaries_U_I_Low_Rifleman {
     displayName = "Crewman";
     role = "Crewman";
     icon = "iconManEngineer";
     engineer = 1;
     weapons[] = {
-        "TACU_Mercenaries_W_HK416_Low",
-        "hgun_Pistol_heavy_01_F",
+        "TACU_Mercenaries_W_Ctar21_Low",
+        "hgun_Pistol_heavy_01_green_F",
         "Throw", "Put"
     };
     respawnWeapons[] = {
-        "TACU_Mercenaries_W_HK416_Low",
-        "hgun_Pistol_heavy_01_F",
+        "TACU_Mercenaries_W_Ctar21_Low",
+        "hgun_Pistol_heavy_01_green_F",
         "Throw", "Put"
     };
-    //EDITORPREVIEW(TACU_Mercenaries_U_O_Low_Crewman);
+    //EDITORPREVIEW(TACU_Mercenaries_U_I_Low_Crewman);
 };
