+++ addons/mercenaries/independent/CfgVehicles_I_Medium.hpp
